pyproject.toml
[project]
name = "snaphelp"
version = "0.1.0"
description = "Assist Marvel Snap players by describing game states and generating strategic advice."
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.5.0",
    "black>=24.0.0",
]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "UP", "B", "ANN", "A"]
ignore = ["ANN101", "ANN102"]

